requirements.txt
tensorflow==2.9.0
pandas==1.4.2
pip==22.1.2
wget==3.2
google-cloud-bigquery==2.34.3
pandas-gbq==0.17.6